Moving opt-in custom signal handler from UnitTests into RCCL
