+++ src/include/signals.h
@@ -1,0 +1,12 @@
+/*************************************************************************
+ * Copyright (c) 2022 Advanced Micro Devices, Inc. All rights reserved.
+ *
+ * See LICENSE.txt for license information
+ ************************************************************************/
+
+#ifndef RCCL_SIGNALS_H_
+#define RCCL_SIGNALS_H_
+
+void RegisterSignalHandlers();
+
+#endif
