--- test/common/config.h.in
@@ -1,6 +1,0 @@
-#cmakedefine HAVE_BFD
-#cmakedefine HAVE_DECL_BFD_GET_SECTION_FLAGS
-#cmakedefine HAVE_DECL_BFD_GET_SECTION_VMA
-#cmakedefine HAVE_TWO_ARG_BFD_SECTION_SIZE
-#cmakedefine HAVE_CPLUS_DEMANGLE
-#cmakedefine HAVE_DECL_BASENAME @HAVE_DECL_BASENAME@
